+++ packages/tracker/src/index.tsx
@@ -80,10 +80,10 @@ export { ENABLE_MIXPANEL_TRACKER } from "./constants";
 export { SUCCESSFUL_WALLET_CONNECTION } from "./constants";
 export { UNSUCCESSFUL_WALLET_CONNECTION } from "./constants";
 
-export { CLICK_BUTTON_CONVERT_AXELAR_BASED } from "./constants";
-export { CLICK_BUTTON_CONFIRM_CONVERT_AXELAR_BASED_TX } from "./constants";
-export { SUCCESSFUL_CONVERT_AXELAR_BASED_TX } from "./constants";
-export { UNSUCCESSFUL_CONVERT_AXELAR_BASED_TX } from "./constants";
+export { CLICK_BUTTON_CONVERT } from "./constants";
+export { CLICK_BUTTON_CONFIRM_CONVERT_TX } from "./constants";
+export { SUCCESSFUL_CONVERT_TX } from "./constants";
+export { UNSUCCESSFUL_CONVERT_TX } from "./constants";
 
 export { CLICK_BUTTON_CONFIRM_WRAP_TX } from "./constants";
 export { SUCCESSFUL_WRAP_TX } from "./constants";