+++ vitest.workspace.ts
@@ -1,4 +1,1 @@
-export default [
-    'apps/*/vitest.config.ts',
-    'packages/*/vitest.config.ts',
-]
+export default ["apps/*/vitest.config.ts", "packages/*/vitest.config.ts"];